backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
requests>=2.28
python-jose[cryptography]>=3.3
python-multipart>=0.0.9
websocket-client>=1.7
